--- lib/chat/userChat.dart
@@ -1,34 +1,0 @@
-class User {
-  final int id;
-  final String name;
-
-  User({
-    this.id,
-    this.name,
-  });
-}
-
-final User currentUser = User(
-  id: 0,
-  name: 'Jose',
-);
-
-final User juancarlos = User(
-  id: 1,
-  name: 'Juan Carlos',
-);
-
-final User cris = User(
-  id: 2,
-  name: 'Cristina',
-);
-
-final User elena = User(
-  id: 3,
-  name: 'Elena',
-);
-
-final User sofi = User(
-  id: 4,
-  name: 'Sofia23',
-);
